Move the conversion table player-type / player-creation to the game setup to avoid having business logic in the menu-options enum.
