+++ src/test/java/com/mael/ttt/GameSetupTest.java
@@ -1,22 +1,30 @@
 package com.mael.ttt;
 
+import com.mael.ttt.players.AlienPlayer;
+import com.mael.ttt.players.HumanPlayer;
+import com.mael.ttt.players.RobotPlayer;
 import com.mael.ttt.ui.SpyConsole;
 import com.mael.ttt.ui.UserInterface;
 import org.junit.Before;
 import org.junit.Test;
 
+import static com.mael.ttt.Mark.PLAYER;
+import static com.mael.ttt.players.PlayerType.*;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
 public class GameSetupTest {
 
     private SpyConsole spy;
     private GameSetup gameSetup;
+    private UserInterface gameUI;
 
     @Before
     public void setUp() {
         int size  = 3;
         spy       = new SpyConsole();
-        gameSetup = new GameSetup(new Board(size), new UserInterface(spy));
+        gameUI    = new UserInterface(spy);
+        gameSetup = new GameSetup(new Board(size), gameUI);
     }
 
     @Test
@@ -38,5 +46,20 @@ public class GameSetupTest {
                       "n");
         gameSetup.playGame();
         assertEquals(UserInterface.REPLAY, spy.lastPrintedMessage());
+    }
+
+    @Test
+    public void createsAHuman() {
+        assertTrue(gameSetup.conversionTable(HUMAN, PLAYER) instanceof HumanPlayer);
+    }
+
+    @Test
+    public void createsARobot() {
+        assertTrue(gameSetup.conversionTable(ROBOT, PLAYER) instanceof RobotPlayer);
+    }
+
+    @Test
+    public void createsAnAlien() {
+        assertTrue(gameSetup.conversionTable(ALIEN, PLAYER) instanceof AlienPlayer);
     }
 }
