+++ src/test/java/com/mael/ttt/ui/MenuOptionTest.java
@@ -1,13 +1,10 @@
 package com.mael.ttt.ui;
 
-import com.mael.ttt.players.AlienPlayer;
-import com.mael.ttt.players.HumanPlayer;
-import com.mael.ttt.players.RobotPlayer;
 import org.junit.Before;
 import org.junit.Test;
 
-import static com.mael.ttt.Mark.PLAYER;
-import static com.mael.ttt.players.PlayerType.*;
+import static com.mael.ttt.players.PlayerType.HUMAN;
+import static com.mael.ttt.players.PlayerType.ROBOT;
 import static com.mael.ttt.ui.MenuOption.HUMAN_HUMAN;
 import static com.mael.ttt.ui.MenuOption.HUMAN_ROBOT;
 import static org.junit.Assert.assertEquals;
@@ -49,20 +46,5 @@ public class MenuOptionTest {
     @Test
     public void returnsHumanVsHumanIfInvalidOption() {
         assertEquals(HUMAN_HUMAN, MenuOption.idToOption("asdfg"));
-    }
-
-    @Test
-    public void createsAHuman() {
-        assertEquals(true, MenuOption.conversionTable(HUMAN, gameUI, PLAYER) instanceof HumanPlayer);
-    }
-
-    @Test
-    public void createsARobot() {
-        assertEquals(true, MenuOption.conversionTable(ROBOT, gameUI, PLAYER) instanceof RobotPlayer);
-    }
-
-    @Test
-    public void createsAnAlien() {
-        assertEquals(true, MenuOption.conversionTable(ALIEN, gameUI, PLAYER) instanceof AlienPlayer);
     }
 }
